application/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
psutil==5.9.6
pydantic==2.5.0 